+++ packages/react-ui/storybook/stories/atoms/IconButton.stories.js
@@ -17,7 +17,7 @@ const options = {
     size: {
       control: {
         type: 'select',
-        options: ['small', 'medium', 'large']
+        options: ['xsmall', 'small', 'medium', 'large']
       }
     },
     disabled: {