+++ packages/react-ui/storybook/stories/atoms/Radio.stories.js
@@ -16,6 +16,11 @@ const options = {
         type: 'boolean'
       }
     },
+    readOnly: {
+      control: {
+        type: 'boolean'
+      }
+    },
     label: {
       control: {
         type: 'text'
@@ -34,12 +39,15 @@ const options = {
 };
 export default options;
 
-const RadioTemplate = ({ label, size, checked, disabled, ...args }) => {
+const RadioTemplate = ({ label, size, checked, disabled, readOnly, ...args }) => {
   return (
     <FormControlLabel
       value={label}
-      control={<Radio size={size} checked={checked} disabled={disabled} />}
+      control={
+        <Radio size={size} checked={checked} disabled={disabled} readOnly={readOnly} />
+      }
       label={label}
+      readOnly={readOnly}
     />
   );
 };
@@ -76,6 +84,25 @@ const StatesTemplate = ({ size, ...args }) => {
           />
         </Grid>
       </Grid>
+
+      <Grid item container spacing={2}>
+        <Grid item xs={4}>
+          <FormControlLabel
+            control={<Radio size={size} checked readOnly />}
+            label='Read Only Active'
+            readOnly
+            {...args}
+          />
+        </Grid>
+        <Grid item xs={4}>
+          <FormControlLabel
+            control={<Radio size={size} readOnly />}
+            label='Read Only Inactive'
+            readOnly
+            {...args}
+          />
+        </Grid>
+      </Grid>
     </Grid>
   );
 };
@@ -102,10 +129,18 @@ const SizeTemplate = ({ ...args }) => {
 
       <Grid item container spacing={2}>
         <Grid item xs={4}>
-          <FormControlLabel control={<Radio checked />} label='Medium Active' {...args} />
+          <FormControlLabel
+            control={<Radio checked size='medium' />}
+            label='Medium Active'
+            {...args}
+          />
         </Grid>
         <Grid item xs={4}>
-          <FormControlLabel control={<Radio />} label='Medium Inactive' {...args} />
+          <FormControlLabel
+            control={<Radio size='medium' />}
+            label='Medium Inactive'
+            {...args}
+          />
         </Grid>
       </Grid>
     </Grid>
@@ -127,7 +162,6 @@ Playground.args = { label: 'Text' };
 
 export const States = StatesTemplate.bind({});
 States.argTypes = disabledStatesArgTypes;
-States.args = { size: 'medium' };
 
 export const Sizes = SizeTemplate.bind({});
 Sizes.argTypes = disabledSizesArgTypes;